+++ jinka-admin/src/services/projects.js
@@ -1,34 +1,35 @@
 import api from './api';
 
 // Project service
+// NOTE: Backend has projects table but endpoints need to be added to adminRoutes.js
 export const projectService = {
     // Get all projects
     getAll: async (params = {}) => {
-        const response = await api.get('/projects', { params });
+        const response = await api.get('/admin/projects', { params });
         return response;
     },
 
     // Get single project
     getOne: async (id) => {
-        const response = await api.get(`/projects/${id}`);
+        const response = await api.get(`/admin/projects/${id}`);
         return response;
     },
 
     // Create project
     create: async (data) => {
-        const response = await api.post('/projects', data);
+        const response = await api.post('/admin/projects', data);
         return response;
     },
 
     // Update project
     update: async (id, data) => {
-        const response = await api.put(`/projects/${id}`, data);
+        const response = await api.put(`/admin/projects/${id}`, data);
         return response;
     },
 
     // Update project progress
     updateProgress: async (id, progress, status) => {
-        const response = await api.patch(`/projects/${id}/progress`, {
+        const response = await api.patch(`/admin/projects/${id}/progress`, {
             progress,
             status,
         });
@@ -37,7 +38,7 @@ export const projectService = {
 
     // Delete project
     delete: async (id) => {
-        const response = await api.delete(`/projects/${id}`);
+        const response = await api.delete(`/admin/projects/${id}`);
         return response;
     },
 };
